requirements/base.txt
Django==1.6.2
bpython==0.12
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
South==0.8.4
psycopg2==2.5.2
django-authtools=0.2.1
celery==3.1.10
django-celery=3.1.9
kombu==3.0.14